pyproject.toml
[project]
name = "media-op"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dashscope>=1.25.0",
    "dotenv>=0.9.9",
    "hydra-core>=1.3.2",
    "jinja2>=3.1.6",
    "langchain==0.3.27",
    "langchain-community>=0.3.31",
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "pymysql>=1.1.2",
    "sqlalchemy>=2.0.44",
    "wxautox>=39.1.42",
]
